re-design the "favorites" page
Trying to make the ‘favorites’ page match the home page, working on
making the table layout work with text and images. I can’t seem to get
the titles and paragraphs to line up with the top of the cell. Did a
lot of googling to figure out how to make cells different widths.